+++ src/Riverline/DynamoDB/Repeater.php
@@ -21,27 +21,92 @@ class Repeater
     }
 
     /**
-     * @param Context\BatchGet $context
-     * @return Collection
+     * @param string $table
+     * @param Context\Scan $context
+     * @return \Riverline\DynamoDB\Collection
      */
-    public function BatchGet(Context\BatchGet $context)
+    public function scan($table, Context\Scan $context = null)
     {
-        $items = new Collection();
+        $collection = new Collection();
         do {
             try {
-                $result = $this->connection->batchGet($context);
-                foreach ($result as $item) {
-                    $items->add($item);
+                $items = $this->connection->scan($table, $context);
+                $collection->merge($items);
+                if ($items->more()) {
+                    $context = $items->getNextContext();
+                } else {
+                    // End
+                    break;
                 }
             } catch (\Riverline\DynamoDB\Exception\ProvisionedThroughputExceededException $e) {
                 // Continue
             }
-        } while($result->more());
+        } while(true);
 
-        return $items;
+        return $collection;
     }
 
-    public function BatchWrite(Context\BatchWrite $context)
+    /**
+     * @param string $table
+     * @param mixed $hash
+     * @param Context\Query $context
+     * @return \Riverline\DynamoDB\Collection
+     */
+    public function query($table, $hash, Context\Query $context = null)
+    {
+        $collection = new Collection();
+        do {
+            try {
+                $items = $this->connection->query($table, $context);
+                $collection->merge($items);
+                if ($items->more()) {
+                    $context = $items->getNextContext();
+                } else {
+                    // End
+                    break;
+                }
+            } catch (\Riverline\DynamoDB\Exception\ProvisionedThroughputExceededException $e) {
+                // Continue
+            }
+        } while(true);
+
+        return $collection;
+    }
+
+    /**
+     * @param Context\BatchGet $context
+     * @return Collection
+     */
+    public function batchGet(Context\BatchGet $context)
+    {
+        $batchCollection = new \Riverline\DynamoDB\Batch\BatchCollection();
+        do {
+            try {
+                $itemsByTable = $this->connection->batchGet($context);
+                foreach ($itemsByTable as $table => $items) {
+                    /** @var $collection Collection */
+                    $collection = $batchCollection[$table];
+
+                    if (null === $collection) {
+                        $batchCollection[$table] = $items;
+                    } else {
+                        $collection->merge($items);
+                        $batchCollection[$table] = $collection;
+                    }
+                }
+                if ($itemsByTable->more()) {
+                    $context = $itemsByTable->getUnprocessedKeysContext();
+                } else {
+                    // End
+                    return $batchCollection;
+                }
+            } catch (\Riverline\DynamoDB\Exception\ProvisionedThroughputExceededException $e) {
+                // Continue
+            }
+        } while(true);
+    }
+
+    public function batchWrite(Context\BatchWrite $context)
     {
         do {
             try {